# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chatterbox TTS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tts/voices": {
      "get": {
        "summary": "List Voices",
        "operationId": "list_voices_v1_tts_voices_get",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Engine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoicesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tts/engine_schema": {
      "get": {
        "summary": "Get Engine Schema",
        "operationId": "get_engine_schema_v1_tts_engine_schema_get",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tts/models": {
      "get": {
        "summary": "List Models",
        "operationId": "list_models_v1_tts_models_get",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tts/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_v1_tts_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectionOptions": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "chunk_marker": {
            "type": "string",
            "title": "Chunk Marker",
            "default": "[[CHUNK]]"
          }
        },
        "type": "object",
        "title": "DirectionOptions"
      },
      "EditOptions": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "trim_silence": {
            "type": "boolean",
            "title": "Trim Silence",
            "default": true
          },
          "normalize": {
            "type": "boolean",
            "title": "Normalize",
            "default": true
          },
          "target_dbfs": {
            "type": "number",
            "title": "Target Dbfs",
            "default": -1.0
          }
        },
        "type": "object",
        "title": "EditOptions"
      },
...
      "EngineSchemaField": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "help": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help"
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "step": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "choices": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choices"
          },
          "visible_if": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible If"
          },
          "serialize_scope": {
            "type": "string",
            "title": "Serialize Scope",
            "default": "engine"
          }
        },
        "type": "object",
        "required": [
          "key",
          "type"
        ],
        "title": "EngineSchemaField"
      },
      "EngineSchemaResponse": {
        "properties": {
          "engine_id": {
            "type": "string",
            "title": "Engine Id"
          },
          "backend_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend Id"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/EngineSchemaField"
            },
            "type": "array",
            "title": "Fields"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "engine_id",
          "capabilities",
          "fields"
        ],
        "title": "EngineSchemaResponse"
      },
...
      "ExportOptions": {
        "properties": {
          "format": {
            "type": "string",
            "const": "wav",
            "title": "Format",
            "default": "wav"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "include_timestamp": {
            "type": "boolean",
            "title": "Include Timestamp",
            "default": true
          },
          "include_model": {
            "type": "boolean",
            "title": "Include Model",
            "default": false
          }
        },
        "type": "object",
        "title": "ExportOptions"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreateResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobCreateResponse"
      },
...
      "ModelInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "ModelInfo"
      },
      "ModelsResponse": {
        "properties": {
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "engine",
          "models"
        ],
        "title": "ModelsResponse"
      },
...
      "TTSJobRequest": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectionOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "export": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "editing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EditOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "engine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Id"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "text_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "raw",
                  "adjusted",
                  "interpreted",
                  "snapshot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Source"
          },
          "text_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Raw"
          },
          "text_adjusted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Adjusted"
          },
          "text_interpreted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Interpreted"
          },
          "text_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Snapshot"
          },
          "chunk_markers": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Markers"
          },
          "engine_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Params"
          },
          "post_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Params"
          },
          "edit_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edit Params"
          }
        },
        "type": "object",
        "title": "TTSJobRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "VoiceInfo"
      },
      "VoicesResponse": {
        "properties": {
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "voices": {
            "items": {
              "$ref": "#/components/schemas/VoiceInfo"
            },
            "type": "array",
            "title": "Voices"
          }
        },
        "type": "object",
        "required": [
          "engine",
          "voices"
        ],
        "title": "VoicesResponse"
      }
    }
  }
}